system fields: relations improvements
* Fixes issue with related records not being checked for existence.

* Added caching of relations to avoid querying.

* Improve code readability by splitting code into separate files, align
  naming in different methods and refactor some similar methods to
  share more code.

* Added dumping of "v@" key for optimistic concurrency control and
  detecting outdated records.
